extract localization options to common file
